Fix kovan boosted pool (#2073)
* Minor uncommited lock change

* Conditionally access decimals
If tokenMeta (e.g. tokenlists) don't contain the wrapped/main tokens then this will error out. On Kovan we don't have these tokens in the tokenllists so it was failing here. It's fine to continue if the tokens aren't in the tokenlists since the liquidity will just default to the subgraph value.