Use updated PoolsRepositories that take a single query
- Refactored the PoolsRepositories in the SDK to take a query on
construction and then fetch only takes a limit and skip. This allows us
to cache the pools so that we can hold many in the repository and
retrieve 10 at a time for usePoolsQuery. This was done so that
first/skip are both numbers to be consistent with the subgraph
repository and so that the Balancer Repository always returns 10 results
instead of sometimes filtering some out.