fix: Fixes a variety of small bugs in the join/exit flows (#2646)
* fix: Refetch balances on join tx success
We do this in useWeb3Watchers using blocknative but the txConfirmed event from blocknative is triggered to quick and so refetching balances doesn't capture the new liquidity position.

* fix: Refetch balances on withdraw tx success
We do this in useWeb3Watchers using blocknative but the txConfirmed event from blocknative is triggered to quick and so refetching balances doesn't capture the new liquidity position.

* fix: Fetch latest onchain pool data before calc
For this particular calculation we rely on accurate pool total supply and token balance data. We previously used the global pool object, expecting the onchain attribute to be up to date but we are not refreshing that anywhere else on entering the withdrawal form. This is a quick hack to fix this local issue with the expectation that:
1. useWithdrawMath will be made redundant as we move to the exit pool provider for all pool types.
2. We refactor the pool type and make it easy to quickly fetch the relevant onchain data when we enter critical flows or things change.

* fix: Make priceImpact static in preview