Improve Token Search Speed
The tokens list started to become very slow since adding many new
prices from the new API because it's doing a O^2 search for token
prices and balances looping over two arrays.

This was done because we compare upper and lower case token addresses.

Instead of doing this, change all addresses to sanitized case (through
getAddress) and use a map to search for prices / balances.