Remove tests that expect usePoolQuery to be loading
Since removing the pool address token injection in usePoolQuery it no
longer waits for injectTokens becuase there are no new tokens to
inject. This causes the query to return instantly in tests which makes
the useLock test fail becuase it expects it to take some time to load.

This removes the isLoading check and just makes the tests wait for it to
be loaded, the isLoading was irrelevant to the actual execution of the
tests.