Fix allowance validation when staking in gauges (#3595)
* Revert "Hotfix - remove post validation"

* Fix approving BPT spending with Gauges

When staking a user must allow a gauge to spend their token. However
when we check for user allowances we only check for tokens the vault
can spend, not the allowances for gauges. By allowing injecting a
spender we can inject the gauge into the spenders list which then
fetches allowances for it when then makes it able to correctly tell if
the user has allowed the gauge to spend their BPT.

This fixes the post approval check always failing when staking BPT.

* Fake injectSpender function to avoid onchain calls during unit testing

* fix: Don't fetch allowances twice

---------

Co-authored-by: Alberto Gualis <alberto@balancer.finance>
Co-authored-by: Gareth Fuller <gareth.a.f@gmail.com>