Show Branded Redirect per Pool Type instead of individual pool ID's
There are a number of Xave / Gyro pools on Polygon that aren't showing
the branded redirect banner because we forgot to add their ID's to the
list. As these pools are always of a certain type it's easier to make
the banner show based on pool type rather than ID's so they all show the
correct banner.