Set mainIndex to 0 if undefined
Some sub-pools like Phantom Stable don't have a mainIndex. In the
subgraph this is set to null (even though it's not allowed). In our code
for flattening token trees we check if a pool is a subPool by if
mainIndex is defined or not, and so if this is left undefined it crashes
as it thinks a subpool is a high level pool.